Major changes for applications and equipment.
Add the notion of application specification keys to allow administrators
to specify the names of specification items that can be used on an
application. This aims to homogenize specifications and to later provide
a sort of checklist when creating a new application for a customer.

Added groups for partner applications (user and admin). Administrators
have the right to create specification keys while users can only read
them (and use them to create specifications on an application).

Various view improvements across the affected modules.